requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.1
generics==2.0.0
jsonify==0.5
permission==0.4.1
Pillow==7.2.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2020.1
sqlparse==0.3.1
